feat: reuse progress indicators on chars page
