+++ apps/mottu_marvel/lib/modules/characters/presentation/widgets/characters_page_progress_indicator.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/widgets.dart';
+import 'package:mottu_design_system/mottu_design_system.dart';
+
+class CharacterPageProgressIndicator extends StatelessWidget {
+  const CharacterPageProgressIndicator({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const SliverFillRemaining(
+      hasScrollBody: false,
+      child: Center(
+        child: Padding(
+          padding: EdgeInsets.only(bottom: 40.0),
+          child: MottuProgressIndicator(
+            size: 30.0,
+          ),
+        ),
+      ),
+    );
+  }
+}
