+++ packages/mottu_design_system/lib/src/atomic/atoms/mottu_progress_indicator.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:mottu_design_system/src/themes/mottu_color_theme.dart';
 
 class MottuProgressIndicator extends StatelessWidget {
   const MottuProgressIndicator({
@@ -12,11 +13,13 @@ class MottuProgressIndicator extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final colors = Theme.of(context).extension<MottuColorsTheme>()!;
+
     return SizedBox(
       width: size,
       height: size,
       child: CircularProgressIndicator(
-        valueColor: AlwaysStoppedAnimation(color),
+        valueColor: AlwaysStoppedAnimation(color ?? colors.backgroundSecondary),
       ),
     );
   }
